Split db conn/creation into a separate file
